// app/dash/src/app/(app)/manage-account/business/company-details/types.ts
export interface ICompanyDetails {
  name: string;
  tradingName: string;
  legalEntity: string;
  companyType: string;
  regNumber: string;
  otherTradingName: string;
  countryOfRegion: string;
  dateOfRegistration: string;
  dateOfIncorporation: string;
  taxVat: string;
  otherContact: string;
  website: string;
  statutory: string;
  publicly: string;
  whereListed: string;
  whyMarket: string;
  nameOfRegulator: string;
}